Refactor `CosIdIntervalShardingAlgorithmTest` to improve readability.
